+++ nativelink-store/src/grpc_store.rs
@@ -449,8 +449,8 @@ impl GrpcStore {
         &self,
         digest: DigestInfo,
         writer: &mut DropCloserWriteHalf,
-        offset: usize,
-        length: Option<usize>,
+        offset: u64,
+        length: Option<u64>,
     ) -> Result<(), Error> {
         let action_result = self
             .get_action_result_from_digest(digest)
@@ -510,7 +510,7 @@ impl StoreDriver for GrpcStore {
     async fn has_with_results(
         self: Pin<&Self>,
         keys: &[StoreKey<'_>],
-        results: &mut [Option<usize>],
+        results: &mut [Option<u64>],
     ) -> Result<(), Error> {
         if matches!(self.store_type, nativelink_config::stores::StoreType::ac) {
             keys.iter()
@@ -521,7 +521,7 @@ impl StoreDriver for GrpcStore {
                     // hope that we detect incorrect usage.
                     self.get_action_result_from_digest(key.borrow().into_digest())
                         .await?;
-                    *result = Some(usize::MAX);
+                    *result = Some(u64::MAX);
                     Ok::<_, Error>(())
                 })
                 .collect::<FuturesUnordered<_>>()
@@ -565,7 +565,7 @@ impl StoreDriver for GrpcStore {
         {
             match missing_digests.binary_search(&digest) {
                 Ok(_) => *result = None,
-                Err(_) => *result = Some(usize::try_from(digest.size_bytes())?),
+                Err(_) => *result = Some(u64::try_from(digest.size_bytes())?),
             }
         }
 
@@ -655,8 +655,8 @@ impl StoreDriver for GrpcStore {
         self: Pin<&Self>,
         key: StoreKey<'_>,
         writer: &mut DropCloserWriteHalf,
-        offset: usize,
-        length: Option<usize>,
+        offset: u64,
+        length: Option<u64>,
     ) -> Result<(), Error> {
         let digest = key.into_digest();
         if matches!(self.store_type, nativelink_config::stores::StoreType::ac) {
